refac: change inventory to module and move tests to module
