+++ Backend/4inventory-api/src/main/java/oliyan_debug/inventory_api/core/security/CorsConfig.java
