+++ Backend/4inventory-api/src/main/java/oliyan_debug/inventory_api/modules/backup/BackupController.java
@@ -4,8 +4,8 @@ import oliyan_debug.inventory_api.modules.category.CategoryRepository;
 import oliyan_debug.inventory_api.modules.category.CategoryService;
 import oliyan_debug.inventory_api.modules.registry.RegistryRepository;
 import oliyan_debug.inventory_api.modules.user.UserRepository;
-import oliyan_debug.inventory_api.repositories.InventoryRepository;
-import oliyan_debug.inventory_api.services.InventoryService;
+import oliyan_debug.inventory_api.modules.inventory.InventoryRepository;
+import oliyan_debug.inventory_api.modules.inventory.InventoryService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
