+++ Backend/4inventory-api/src/main/java/oliyan_debug/inventory_api/modules/category/CategoryController.java
@@ -1,6 +1,6 @@
 package oliyan_debug.inventory_api.modules.category;
 
-import oliyan_debug.inventory_api.domain.records.MessageHandler;
+import oliyan_debug.inventory_api.core.utils.MessageHandler;
 import oliyan_debug.inventory_api.modules.category.exceptions.CategoryAlreadyExistsException;
 import oliyan_debug.inventory_api.modules.category.exceptions.CategoryIdNotFoundException;
 import oliyan_debug.inventory_api.modules.category.exceptions.InvalidCategoryException;
