+++ Backend/4inventory-api/src/main/java/oliyan_debug/inventory_api/modules/inventory/InventoryController.java
@@ -1,14 +1,13 @@
-package oliyan_debug.inventory_api.controllers;
+package oliyan_debug.inventory_api.modules.inventory;
 
-import oliyan_debug.inventory_api.domain.entities.InventoryItem;
-import oliyan_debug.inventory_api.domain.exceptions.items.ItemIdNotFoundException;
-import oliyan_debug.inventory_api.domain.exceptions.items.InvalidItemNameException;
-import oliyan_debug.inventory_api.domain.exceptions.items.InvalidQuantityException;
-import oliyan_debug.inventory_api.domain.records.MessageHandler;
+import oliyan_debug.inventory_api.core.utils.MessageHandler;
+import oliyan_debug.inventory_api.modules.inventory.exceptions.InvalidItemNameException;
+import oliyan_debug.inventory_api.modules.inventory.exceptions.InvalidQuantityException;
+import oliyan_debug.inventory_api.modules.inventory.exceptions.ItemIdNotFoundException;
 import oliyan_debug.inventory_api.modules.registry.ItemAndRegistryDTO;
 import oliyan_debug.inventory_api.modules.registry.ItemDelete;
 import oliyan_debug.inventory_api.modules.registry.exceptions.JustificationNotFoundException;
-import oliyan_debug.inventory_api.services.InventoryService;
+import oliyan_debug.inventory_api.modules.inventory.InventoryService;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
