+++ Backend/4inventory-api/src/main/java/oliyan_debug/inventory_api/modules/inventory/ItemAndCategory.java
@@ -1,4 +1,4 @@
-package oliyan_debug.inventory_api.domain.records;
+package oliyan_debug.inventory_api.modules.inventory;
 
 public record ItemAndCategory(Long itemId, Long categoryId) {
 }
