+++ Backend/4inventory-api/src/main/java/oliyan_debug/inventory_api/modules/search/SearchService.java
@@ -1,7 +1,7 @@
 package oliyan_debug.inventory_api.modules.search;
 
-import oliyan_debug.inventory_api.domain.entities.InventoryItem;
-import oliyan_debug.inventory_api.repositories.InventoryRepository;
+import oliyan_debug.inventory_api.modules.inventory.InventoryItem;
+import oliyan_debug.inventory_api.modules.inventory.InventoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
