+++ Backend/4inventory-api/src/test/java/oliyan_debug/inventory_api/modules/inventory/InventoryServiceTest.java
@@ -1,15 +1,16 @@
-package oliyan_debug.inventory_api.services;
+package oliyan_debug.inventory_api.modules.inventory;
 
-import oliyan_debug.inventory_api.domain.entities.InventoryItem;
-import oliyan_debug.inventory_api.domain.exceptions.items.InvalidItemNameException;
-import oliyan_debug.inventory_api.domain.exceptions.items.InvalidQuantityException;
-import oliyan_debug.inventory_api.domain.exceptions.items.ItemIdNotFoundException;
+import oliyan_debug.inventory_api.modules.inventory.InventoryCategoryService;
+import oliyan_debug.inventory_api.modules.inventory.InventoryItem;
+import oliyan_debug.inventory_api.modules.inventory.exceptions.InvalidItemNameException;
+import oliyan_debug.inventory_api.modules.inventory.exceptions.InvalidQuantityException;
+import oliyan_debug.inventory_api.modules.inventory.exceptions.ItemIdNotFoundException;
 import oliyan_debug.inventory_api.modules.registry.ItemAndRegistryDTO;
 import oliyan_debug.inventory_api.modules.registry.ItemDelete;
 import oliyan_debug.inventory_api.modules.registry.RegistryRepository;
 import oliyan_debug.inventory_api.modules.registry.exceptions.JustificationNotFoundException;
-import oliyan_debug.inventory_api.repositories.InventoryRepository;
-import jakarta.persistence.criteria.CriteriaBuilder;
+import oliyan_debug.inventory_api.modules.inventory.InventoryRepository;
+import oliyan_debug.inventory_api.modules.inventory.InventoryService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
