+++ Backend/4inventory-api/src/test/java/oliyan_debug/inventory_api/modules/search/SearchServiceTest.java
@@ -1,8 +1,8 @@
-package oliyan_debug.inventory_api.services;
+package oliyan_debug.inventory_api.modules.search;
 
-import oliyan_debug.inventory_api.domain.entities.InventoryItem;
+import oliyan_debug.inventory_api.modules.inventory.InventoryItem;
 import oliyan_debug.inventory_api.modules.search.SearchService;
-import oliyan_debug.inventory_api.repositories.InventoryRepository;
+import oliyan_debug.inventory_api.modules.inventory.InventoryRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
